Revert "Hg/work on game"
