--- src/components/Square/index.js
@@ -1,16 +1,0 @@
-import React from 'react';
-
-const Square = ({ value, onClick }) => {
-
-    return (
-        <button
-            className="square"
-            onClick={onClick}
-        >
-            {value}
-        </button>
-    );
-
-};
-
-export default Square;
